[Tooling] Combine some tools to a single binary.
