+++ crates/aptos-debugger/src/lib.rs
@@ -1,0 +1,31 @@
+// Copyright © Aptos Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+use anyhow::Result;
+use clap::Parser;
+
+#[derive(Parser)]
+pub enum Cmd {
+    #[clap(subcommand)]
+    AptosDb(aptos_db_tool::DBTool),
+
+    Decode(aptos_move_debugger::bcs_txn_decoder::Command),
+
+    Move(aptos_move_debugger::aptos_debugger::Command),
+}
+
+impl Cmd {
+    pub async fn run(self) -> Result<()> {
+        match self {
+            Cmd::AptosDb(cmd) => cmd.run().await,
+            Cmd::Decode(cmd) => cmd.run().await,
+            Cmd::Move(cmd) => cmd.run().await,
+        }
+    }
+}
+
+#[test]
+fn verify_tool() {
+    use clap::CommandFactory;
+    Cmd::command().debug_assert()
+}
