+++ crates/aptos-debugger/src/main.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use anyhow::Result;
-use aptos_db_tool::DBTool;
+use aptos_debugger::Cmd;
 use aptos_logger::{Level, Logger};
 use aptos_push_metrics::MetricsPusher;
 use clap::Parser;
@@ -12,6 +12,5 @@ async fn main() -> Result<()> {
     Logger::new().level(Level::Info).init();
     let _mp = MetricsPusher::start(vec![]);
 
-    DBTool::parse().run().await?;
-    Ok(())
+    Cmd::parse().run().await
 }
