+++ docker/builder/validator.Dockerfile
@@ -28,8 +28,7 @@ RUN addgroup --system --gid 6180 aptos && adduser --system --ingroup aptos --no-
 
 RUN mkdir -p /opt/aptos/etc
 COPY --link --from=node-builder /aptos/dist/aptos-node /usr/local/bin/
-COPY --link --from=tools-builder /aptos/dist/aptos-db-tool /usr/local/bin/
-COPY --link --from=tools-builder /aptos/dist/aptos-db-bootstrapper /usr/local/bin/
+COPY --link --from=tools-builder /aptos/dist/aptos-debugger /usr/local/bin/
 
 # Admission control
 EXPOSE 8000
