+++ storage/db-tool/Cargo.toml
@@ -17,6 +17,7 @@ aptos-backup-cli = { workspace = true }
 aptos-backup-service = { workspace = true }
 aptos-config = { workspace = true }
 aptos-db = { workspace = true, features = ["db-debugger"] }
+aptos-executor = { workspace = true }
 aptos-executor-types = { workspace = true }
 aptos-logger = { workspace = true }
 aptos-push-metrics = { workspace = true }
@@ -24,7 +25,9 @@ aptos-state-view = { workspace = true }
 aptos-storage-interface = { workspace = true }
 aptos-temppath = { workspace = true }
 aptos-types = { workspace = true }
+aptos-vm = { workspace = true }
 async-trait = { workspace = true }
+bcs = { workspace = true }
 clap = { workspace = true }
 itertools = { workspace = true }
 owo-colors = { workspace = true }
