+++ testsuite/module_verify_run_local.py
@@ -29,7 +29,7 @@ def local_setup():
             "build",
             "--release",
             "-p",
-            "aptos-db-tool",
+            "aptos-debugger",
         ],
         check=True,
     )
