+++ testsuite/smoke-test/Cargo.toml
@@ -20,7 +20,6 @@ aptos-cached-packages = { workspace = true }
 aptos-config = { workspace = true }
 aptos-consensus = { workspace = true }
 aptos-crypto = { workspace = true }
-aptos-debugger = { workspace = true }
 aptos-faucet-core = { workspace = true }
 aptos-forge = { workspace = true }
 aptos-framework = { workspace = true }
@@ -30,6 +29,7 @@ aptos-global-constants = { workspace = true }
 aptos-indexer = { workspace = true }
 aptos-inspection-service = { workspace = true }
 aptos-keygen = { workspace = true }
+aptos-move-debugger = { workspace = true }
 aptos-release-builder = { workspace = true }
 aptos-rest-client = { workspace = true }
 aptos-rosetta = { workspace = true }
