+++ testsuite/smoke-test/src/aptos/mint_transfer.rs
@@ -3,8 +3,8 @@
 
 use crate::smoke_test_environment::new_local_swarm_with_aptos;
 use aptos_cached_packages::aptos_stdlib;
-use aptos_debugger::AptosDebugger;
 use aptos_forge::Swarm;
+use aptos_move_debugger::aptos_debugger::AptosDebugger;
 use aptos_types::transaction::{ExecutionStatus, TransactionStatus};
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
