+++ testsuite/verify_core/common.py
@@ -43,7 +43,8 @@ def query_backup_latest_version(backup_config_template_path: str) -> int:
     """query latest version in backup, at the same time, pre-heat metadata cache"""
     db_backup_result = subprocess.Popen(
         [
-            "target/release/aptos-db-tool",
+            "target/release/aptos-debugger",
+            "aptos-db",
             "backup",
             "query",
             "backup-storage-state",
